+++ assetsAMD/js/views/bookdetail.js
@@ -1,6 +1,6 @@
 // bookdetail.js: Ted Killilea June 2012  @t2k_nyc
 // templates/bookdetail.htm
-define(['underscoreM',  'marionette', 'templates' ], function(_, Marionette, templates) {
+define(['underscoreM', 'marionette', 'templates' ], function(_, Marionette, templates) {
     'use strict';
 
     return Marionette.ItemView.extend({